pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sheaia"
version = "0.1.0"
description = "AI-Powered Enterprise Insights Platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "SHEAIA Team" }
]
keywords = ["ai", "enterprise", "langgraph", "llm", "data-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # LLM & AI
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "langmem>=0.0.10",
    
    # LLM Serving
    "llama-cpp-python>=0.3.0",
    
    # Embeddings
    "sentence-transformers>=3.0.0",
    
    # Vector DB
    "pymilvus>=2.4.0",
    
    # Database
    "clickhouse-connect>=0.7.0",
    "sqlalchemy>=2.0.0",
    
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",
    
    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.0",  # Excel support
    "pypdf>=4.0.0",     # PDF support
    "python-docx>=1.1.0",  # Word support
    
    # Database connectors
    "pyodbc>=5.1.0",
    
    # Utilities
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "networkx>=3.3",  # Schema graph
    
    # Security
    "cryptography>=43.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
wechat = [
    "wechatpy>=1.8.0",
]
feishu = [
    "feishu-python-sdk>=0.1.0",
]

[project.scripts]
sheaia = "sheaia.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/sheaia"
Documentation = "https://github.com/your-org/sheaia#readme"
Repository = "https://github.com/your-org/sheaia"

[tool.hatch.build.targets.wheel]
packages = ["sheaia"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
